reducing size of main app entry point
